処理個別のメソッドを削除して、runとgetRecordのみでクエリを実行するように修正
